backend/requirements.txt
asgiref==3.8.1
attrs==24.3.0
cffi==1.17.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
psycopg2-binary==2.9.10
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.22.3
sqlparse==0.5.3
uritemplate==4.1.1
